Improves development experience and project setup
Enhances the development workflow by automatically opening the application in the browser when `ng serve` is run.
Also, configures the build process to watch for changes during development.

Adds placeholder README files for audio and image assets to guide developers on where to place these resources.

Updates the application title and adds meta description and a Google Fonts link to the `index.html` for improved SEO and aesthetics.